+++ packages/core/styles/src/properties/base.ts
@@ -43,4 +43,5 @@ export const baseColors = {
   "neutral-textDisabled": varsThemeBase.colors.neutral.textDisabled,
   "neutral-textLow": varsThemeBase.colors.neutral.textLow,
   "neutral-textHigh": varsThemeBase.colors.neutral.textHigh,
+  "ai-gradient": varsThemeBase.colors.aiGenerative.interactive,
 };