+++ packages/react/src/atomic/Button/src/button.spec.tsx
@@ -59,12 +59,5 @@ describe("GIVEN <Button />", () => {
         screen.getByRole("button", { name: "button" }).getAttribute("class")
       ).toContain("appearance_transparent");
     });
-
-    it("AND should correctly render the appearance ai-gradient", () => {
-      makeSut({ appearance: "ai-gradient", children: "button" });
-      expect(
-        screen.getByRole("button", { name: "button" }).getAttribute("class")
-      ).toContain("appearance_ai-gradient");
-    });
   });
 });