+++ packages/react/src/atomic/Chip/src/chip.spec.tsx
@@ -34,11 +34,4 @@ describe("GIVEN <Chip />", () => {
       "appearance_neutral"
     );
   });
-
-  it("THEN it should correctly render the ai-gradient appearance", () => {
-    makeSut({ text: "chip", appearance: "ai-gradient" });
-    expect(screen.getByTestId("chip-element").getAttribute("class")).toContain(
-      "appearance_ai-gradient"
-    );
-  });
 });