+++ packages/core/tokens/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 Our design tokens are the foundations of Nimbus Design System.
 
+## 2025-07-29 `9.2.0`
+
+#### 🎉 New features
+
+- Added new `aiGenerative` system color tokens. ([#320](https://github.com/TiendaNube/nimbus-design-system/pull/320) by [@joacotornello](https://github.com/joacotornello))
+
 ## 2025-03-18 `9.1.1`
 
 ### 💡 Others